+++ app/app/Models/GroupPermissionSchema.cs
@@ -5,7 +5,7 @@ namespace app.Models;
 [Table("group_permission_schema")]
 public class GroupPermissionSchema
 {
-  [Column("id_permission")] public int IdPermission { get; set; }
-  [Column("id_schema")] public int IdSchema { get; set; }
-  [Column("id_group_user")] public int IdGroupUser { get; set; }
+  [Column("id_permission")] public long IdPermission { get; set; }
+  [Column("id_schema")] public long IdSchema { get; set; }
+  [Column("id_group_user")] public long IdGroupUser { get; set; }
 }
