+++ app/app/Models/Privacy.cs
@@ -5,6 +5,6 @@ namespace app.Models;
 [Table("privacy")]
 public class Privacy
 {
-  [Column("id_privacy")] public int Id { get; set; }
+  [Column("id_privacy")] public long Id { get; set; }
   [Column("type_privacy")] public string? Type { get; set; }
 }
