+++ app/app/Models/Schema.cs
@@ -5,7 +5,7 @@ namespace app.Models;
 [Table("schema")]
 public class Schema
 {
-  [Column("id_schema")] public int Id { get; set; }
+  [Column("id_schema")] public long Id { get; set; }
   [Column("is_public")] public bool? IsPublic { get; set; }
   [Column("name_schema")] public string? Name { get; set; }
   [Column("desc_schema")] public string? Desc { get; set; }
